requirements.txt
matplotlib>=3.5.1
opencv_python>=4.10.0.84
pillow
torch>=2.4.1
torchvision>=0.19.1
tqdm>=4.66.5
numpy>=1.26.4
peft>=0.14.0
sentencepiece>=0.2.0
scikit-learn>=1.6.1
pandas>=2.2.3
pyarrow>=19.0.0
pytest>=7.4.0
pytorch-forecasting>=1.3.0